fix: Properly handle empty CSVs with valid headers
The Polars `raise_if_empty` flag only errors if the file is completely empty, so the
majority of "empty" files from edx fail to trigger that path since they contain a header
row. This does a follow-up check to see if the resulting data frame is empty of any
data-containing rows.

+++ src/ol_orchestrate/assets/edxorg_archive.py
@@ -257,7 +257,7 @@ def process_edxorg_archive_bundle(
             # (TMM 2024-03-14)
             if table_name:
                 try:
-                    pl.read_csv(
+                    df = pl.read_csv(
                         archive_file,
                         has_header=True,
                         n_rows=2,
@@ -270,6 +270,9 @@ def process_edxorg_archive_bundle(
                     context.log.debug(
                         "Skipping further processing of empty CSV file %s", tinfo.name
                     )
+                    archive_file.unlink()
+                    continue
+                if df.is_empty():
                     archive_file.unlink()
                     continue
             data_version = hashlib.file_digest(
